requirements.txt
APScheduler==3.6.3
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.7.0
pyshorteners==1.0.1
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
